Updates for mongoid2: Updated gemspec. Updated spec mongoid db initialization. Updated timestamp. Removed counter_cache. Updated specs to pass.

+++ lib/mongoid_rating/rate.rb
@@ -1,7 +1,7 @@
 module Mongoid::Rating
   class Rate
     include Mongoid::Document
-    include Mongoid::Timestamps::Short
+    include Mongoid::Timestamps
     
     embedded_in :rateable, polymorphic: true
     belongs_to :rater, polymorphic: true, inverse_of: nil
